% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem2stl.R
\name{ellipsoidnormal}
\alias{ellipsoidnormal}
\title{Calculate the surface normal vector at a specific point on an ellipsoid.}
\usage{
ellipsoidnormal(lat, long, a = 1, b = 1, c = 1)
}
\arguments{
\item{lat}{latitude coordinates (in radians)}

\item{long}{longitude coordinates (in radians)}

\item{a}{semi-principal axis length of the ellipsoid}

\item{b}{semi-principal axis length of the ellipsoid}

\item{c}{semi-principal axis length of the ellipsoid}
}
\description{
Calculate the surface normal vector at a specific point on an ellipsoid.
}
