% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem2stl.R
\name{latlong2ellipsoid}
\alias{latlong2ellipsoid}
\title{Map latitude/longitude coordinates onto an ellipsoid. Vectorised.}
\usage{
latlong2ellipsoid(lat, long, a = 1, b = 1, c = 1)
}
\arguments{
\item{lat}{latitude coordinates (scalar or vector, in radians)}

\item{long}{longitude coordinates (same length as \code{lat}, in radians)}

\item{a}{semi-principal axis length}

\item{b}{semi-principal axis length}

\item{c}{semi-principal axis length}
}
\value{
a \code{length(lat)x3} matrix giving the lat/long coordinates points
  on an ellipsoid with the given axes lengths in 3d space
}
\description{
Map latitude/longitude coordinates onto an ellipsoid. Vectorised.
}
