% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem2stl.R
\name{mesh2stl}
\alias{mesh2stl}
\alias{mesh2obj}
\alias{mesh2ply}
\alias{mesh2webgl}
\title{Write mesh objects to various file formats.}
\usage{
mesh2stl(filename, x, ...)

mesh2obj(filename, x, ...)

mesh2ply(filename, x, ...)

mesh2webgl(filename, x, ...)
}
\arguments{
\item{filename}{the file to write to}

\item{x}{an rgl \code{mesh3d} object}

\item{...}{extra arguments passed to the corresponding
\code{rgl::write...()} function}
}
\value{
Invisibly returns the name of the connection to which the data was
  written.
}
\description{
Write mesh objects to various file formats.
}
