% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem2stl.R
\name{xprod}
\alias{xprod}
\title{Calculate vector crossproduct.}
\usage{
xprod(...)
}
\arguments{
\item{...}{\code{n} vectors, each of length \code{n+1}}
}
\value{
a vector of the same length as the \code{...} arguments
}
\description{
Calculate vector crossproduct.
}
